requirements-tflite.txt
# Requirements for PyTorch to TFLite conversion
# Install with: pip install -r requirements-tflite.txt

# Core PyTorch
torch>=1.13.0
torchvision>=0.14.0

# ONNX
onnx>=1.14.0
onnx-simplifier>=0.4.0

# TensorFlow
tensorflow>=2.12.0,<2.16.0  # Avoid very latest for compatibility

# Conversion tools (choose one)
onnx2tf>=1.16.0  # Recommended for 2025
# onnx-tf>=1.10.0  # Alternative (uncomment if using this instead)

# Additional utilities
numpy>=1.23.0
Pillow>=9.0.0

# Optional: For visualization
# netron>=6.0.0  # View ONNX/TFLite models graphically
